Build nodeset of data producers and give it to JS
The nodeset of nodes that have produced metrics data for the job is
computed by the backend (with the help of the ClusterShell library)
and is sent in the JSON hash to the JS on the frontend side.

Then, the JS displays this information in the HTML DOM.

This requires adding a new dependency on ClusterShell on the backend
REST API package.