New backend profiling mechanism
This commit introduces a new Profiler singleton object in the backend
application. It is used all across the application to handle a set of
timers and debug metadata.

Those profiling/debug information are then sent to the clients in the
response JSON hash.

The HTML/JS then displays those information inside an optional modal
when the debug URL parameter is set.

Fixes #2.